fix bug where the network would just drop circuits
It turned out that by separating the "create the cell that will be sent
when opening a stream" from "send the cell to open a stream", another
stream could send bytes first, which means they would be out of order on
the wire, and so decrypting them would fail. Really what we need to do
here when writing is buffered is to queue all cells to be written and
poll on actually sending them. Circuit probably needs another
refactor/restructure.